requirements.txt
fastapi==0.115.6
Pillow==11.1.0
protobuf==5.29.3
python-dotenv==1.0.1
uvicorn==0.34.0
google-generativeai